# src/main/java/com/javaweb2/dto/WorkerDTO.java
package com.javaweb2.dto;

import java.time.LocalDateTime;

public class WorkerDTO {
    public Long id;

    public Long supplierId;

    public String name;

    public String email;

    public String password;

    public String role;

    public LocalDateTime createdAt;
}